Fix top-bottom margin issue
Driver was sending 33 blank lines at the top of the page and 38 at the bottom
(this averages out to about 35). This combined with the hardware 35-line
margin on die-cut labels caused a ~6mm shift in the vertical alignment on
QL-500 after we fixed the margin bug.

Generally speaking we shouldn't be adding margins unless the user has asked
for them in the printer driver settings.