Set the driver unprintable margins to 0
Without an unprintable margin definition, foomatic will assume unprintable
margins of [18 36 18 36] pt, which is much larger than what the ptouch printers
support.  This will result in incorrect HWMargins and ImageableArea statements
in PPD files, and programs that respect those definitions will not be able to
use the actual printable area.

To avoid that, set the driver margins to 0.  Specific printers can override the
margins in the printer definition.