Bring some sanity to the foomatic printer definitions
The foomatic printer definitions consist of driver, printer, and print option
definitions (<driver>, <printer>, <option>).

There are some redundancies and inefficiencies in these definitions:
* Each driver defines which printers it supports.
* Each printer defines its preferable driver.
* In addition, the option definition includes which driver and/or printer
  combinations that option applies to.  This means that adding support for a
  new printer requires going through all the options the printer supports and
  adding the printer.  With many options, this quickly turns into a huge mess.

To get rid of these problems, we introduce an optional <select> tag to printer
definitions that defines which options that particular printer supports.  With
that, adding a new printer that is similar to an existing printer, as well as
comparing printer definitions, becomes much easier.

To turn these definitions into something foomatic understands, we turn these
<select> tags into the corresponding <constraint> tags in option definitions
using a script called "foomaticalize".

The syntax of <select> tags is described in that script.

Note: The foomatic-db documentation mentions that the list of drivers that
support a printer can also be put into the <printer> definition as
<drivers><driver><id>...</id></driver></drivers>.  However, that didn't work.