rastertoptch: Remove old C compatibility code
We should be safe to assume that headers like <string.h> and <stdbool.h> are
available these days.  Remove all the related autoconf tests and compatibility
code.  If ptouch-driver should still need to be build on systems that don't
provide that functionality, we should use gnulib to emulate the standard
interfaces for us instead of working around it in the code.