rastertoptch: Use new mode switch command on new printers
rastertoptch has so far used "ESC i R 0x01" to switch PT printers into raster
mode.  However, the Brother Raster Command Reference manuals document that the
"ESC i a 0x01" command sequence should be used for that purpose instead, and
that's the command sequence at least the current Windows drivers seem to be
using.  We don't have a way to find out which printers require "ESC i R 0x01",
so keep using that for old printers, but switch to "ESC i a 0x01" for new
ones.