rastertoptch: Add Debug option
Print the job options as well as all the fields from the cups page header(s)
that control the printing of the pages.

To see these messages, you will have to set the LogLevel in cupsd.conf to
debug.  This will make cups extremely chatty, so you may have to search for
"rastertoptch" in the log to find the relevant messages.