rastertoptch: Fix empty lines in NegativePrint
There are three bugs in RLE_store_empty_lines when NegativePrint is enabled:

First, this function will produce incorrect output for all lines after the
first: with bytes_per_line == 16, the first line will be encoded as
"G 02 00 f1 ff", which is correct.  bytes_per_line will now be 0, and all
remaining lines which will be encoded as just "G 02 00", which is incorrect.

Second, a 256-byte line would be encoded as "G 02 00 81 ff" "G 02 00 81 ff"
"G 02 00 02 ff".  This is incorrect because the entire line must be encoded
as a single "G" command.

Third, the code allows at most 128 repetitions, but the largest number of
repetitions that can be represented is 129 (which is encoded as 0x80).  So the
correct encoding for a 256-byte "empty" line is actually "G 04 00 80 ff 82 ff".

Fix those issues.