rastertoptch: Turn postscript page options into cups options
Commands like ESC i M are always per job and not per page, so when such
commands are used, they take effect immediately, even for pages that have been
buffered but haven't been printed, yet.  This means that for getting
controllable results, those commands shouldn't be used after a Form Feed.
(Verified on PT-18R and PT-P700.)

Because of that, it doesn't make much sense to use the CutMedia, AdvanceMedia,
and AdvanceDistance page headers.  Replace those with cups options.  This
matches what the drivers supplied by Brother are doing.

In addition, fix the use of Form Feed (end of page) vs. ^Z (end of job): the
Form Feed command (0x0c) indicates the end of a page, with more pages to be
printed.  The ^Z command (0x1a) indicates the end of the print job.  Some
printers actually expect a ^Z command at the end of the job and when that
doesn't arrive, they keep waiting for more data.