Update option Brother-PT-PageSize
Define the actual sizes of Brother TTZe tapes and HSe Heat Shrink Tubes.  Turn
the supported page sizes into "opt-in" enumeration values and list explicitly
which printers support which page sizes.

The shortname of "TZe Tape (3.5mm)" is "tze-4" because calling it "tze-3.5"
would break the order in which the page sizes appear in PPD files (and
consequently, in the UI).

Until January 31, 2020, the foomatic parser didn't recognize values with
fractional numbers like "9.6 2834.65" in PageSize values.  Once that fix is
available, we can switch to the actual values in the PageSize definitions.