rastertoptch: Remove broken feed code
The code in emit_feed_cut_mirror first computes a five-bit "feed_nibble" in an
unknown unit and then puts the feed amount in pt units in the lower five bits
of the ESC i M command byte.  However, the AdvanceDistance option allows the
largest feed to be 74 pt, which cannot be represented in five bits.  This all
suggests that the code in question has never worked, so remove it.

The lower five bits of the ESC i M command are documented as being
unused in all the reference manuals I could find.  The documented command
for specifying an additional margin (i.e., an extra feeed) is ESC i d, which
was confirmed to work.