Avoid automake noinst_ problems
We always want to build the ptexplain debugging utility to make sure it won't
break, without installing it.  The documented way to achieve that is by using
the noinst_ prefix.  It turns out that noinst_ doesn't always work correctly in
combination with a target name (here, "ptexplain"), but we can omit the target
name in this case.