Add label "recovery" option
It turns out that when the PI_RECOVER flag of the print information command
(ESC i z) is set, PT-P700 will verify if the tape inserted in the printer
matches the print job, and the printer will go into an error state if the tape
doesn't match.  The printer *probably* also sets the Wrong Media flag in its
error codes, but the rastertoptch driver cannot use two-way communication with
the printer, so we unfortunately cannot check that.

It can be frustrating to users if the printer just goes into an error state
with no further information, so add a flag that makes setting the PI_RECOVER
flag an option ("Only print on labels of matching size") that is off by
default for now.

On QL printers, the PI_RECOVER flag does not allow to override the MediaType
check.  The printer will still refuse to print continuous-length tape print
jobs onto die-cut labels and vice versa.