Add LastPageFlag option for PT-P900 series
The page flag of the print information command was so far always set to 0 on
the first page, and 1 on all other pages.  The PT-P900 series are the first
printers that expects the page flag to be 2 on the last page (even if it's also
the first page).

Since this is the uncommon case so far and we don't know if a page flag of 2
will confuse any of the existing printers that use the print information
command (i.e., QL-*, PT-P700, PT-E500W), add a LastPageFlag option to
explicitly enable setting the page flag to 2 for now.