Add a MediaType option for QL printers
Instead of the previous hack of choosing the MediaType based on the PageSize
option which no longer works due to following Ghostscript bug:

  https://bugs.ghostscript.com/show_bug.cgi?id=702111

Turn the MediaType into a separate, user-visible option.  This obsoletes the
RollFedMedia option.

I've explored using a foomatic Composite Option instead, but that's not
possible because foomatic generates the PaperDimension and ImageableArea
definition based on the <ev_driverval> values of the PageSize option, and with
a Composite Option, those values wouldn't contain the page dimensions anymore.
(In addition, we'd need a "Custom Label" and a "Custom Tape" option, which also
isn't supported.)