Add a MinMargin option
As explained in more detail in the code, QL printers require a specific minimum
top and bottom margin to ensure that they can cut properly between labels.
That margin seems to be 3mm on all QL printers.

PT printers have printer-dependent minimum margin amounts.  On those printers,
the minimum margins are not enforced, but they still ensure that labels are cut
in a blank space, so set the minimum margin amounts for those PT printers for
which there's a specification.