Disable the NegativePrint option
No longer offer the NegativePrint option: this option is mostly useful for
debugging the driver, but doesn't actually produce decent looking labels; it
isn't suitable for printing white-on-black labels using black-on-white tape
because there will be narrow but visible margins, for example.

The Brother drivers do not offer this otion.  There could be a risk of
overheating and eventually damaging the print head when trying to print outside
of the tape too much.