Change non-colorized selected bits markers in ptexplain
In a recent commit, the ptexplain output was changed to highlight the
selected bits instead of the deselected bits, which is more intuitive.
However, this also makes the output ambiguous if you do not know whether
the old or new ptexplain was used to generate the output.

This changes the output to use <> instead [] to mark the selected bits,
so the ouput is no lo longer ambiguous.